# api/openapi.yaml
openapi: 3.1.0
info:
  title: CarMarket API
  version: 0.1.0
  description: India used car marketplace REST API
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
paths:
...
  /dealer-stores/me/listings:
    get:
      summary: Dealer store inventory
      tags: [dealer-stores]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Listings
  /listings:
    get:
      summary: Search listings
      tags: [listings]
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Listing search results
    post:
      summary: Create listing
      tags: [listings]
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created
  /listings/me:
    get:
      summary: My listings
      tags: [listings]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Owner listings
  /listings/{id}:
    get:
      summary: Get listing detail
      tags: [listings]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Listing detail
    patch:
      summary: Update listing
      tags: [listings]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Updated
    delete:
      summary: Remove listing
      tags: [listings]
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Removed
  /listings/{id}/publish:
    post:
      summary: Submit listing for review
      tags: [listings]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Published for review
  /listings/{id}/sold:
    post:
      summary: Mark listing as sold
      tags: [listings]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Marked sold
  /listings/{id}/renew:
    post:
      summary: Renew listing expiry
      tags: [listings]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Renewed
  /listings/{id}/duplicate:
    post:
      summary: Duplicate listing as draft
      tags: [listings]
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Duplicated
  /listings/{id}/images/presign:
    post:
      summary: Get presigned upload URL
      tags: [listings]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Presigned URL
  /listings/{id}/images/confirm:
    post:
      summary: Confirm image upload
      tags: [listings]
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Image recorded
  /listings/{id}/images/{image_id}:
    delete:
      summary: Delete listing image
      tags: [listings]
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Deleted
...
  /reports/listings/{listing_id}:
    post:
      summary: Report a listing
      tags: [reports]
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Report filed
...
  /moderation/listings:
    get:
      summary: Pending moderation queue
      tags: [moderation]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Queue
  /moderation/listings/{id}/approve:
    post:
      summary: Approve listing
      tags: [moderation]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Approved
  /moderation/listings/{id}/reject:
    post:
      summary: Reject listing
      tags: [moderation]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Rejected
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT